+++ CHANGES.rst
@@ -24,6 +24,8 @@
 - Use randomized data for the objects instead of a constant string.
   This lets us more accurately model effects due to compression at the
   storage or network layers.
+- When gevent is installed, add support for testing with the system
+  monkey patched. (Note: This might not be supported by all storages.)
 - Add ``--zap`` to rebuild RelStorage schemas.
 
 0.5 (2012-09-08)